+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+-   Support for block energy limit chain update.
+
 ## 1.5.0
 
 ### Added